refactor(protondb-search): extract constants
